POS Modularization: Move SwiftUI view modifiers to WooFoundation
- Move View+AutofocusTextModifier to WooFoundation for cross-module reuse
- Move View+Conditionals to WooFoundation for cross-module reuse
- Move View+Measurements to WooFoundation for cross-module reuse
- Move View+RoundedBorder to WooFoundation for cross-module reuse
- Move View+ScrollModifiers to WooFoundation for cross-module reuse
- Move View+SizeTracker to WooFoundation for cross-module reuse
