+++ Modules/Sources/WooFoundation/ViewModifiers/View+Conditionals.swift
@@ -1,6 +1,6 @@
 import SwiftUI
 
-extension View {
+public extension View {
     /// Renders a view if the provided  `condition` is met.
     /// If the `condition` is not met, an `nil`  will be used in place of the receiver view.
     ///
